requirements.txt
torch==2.0.1
numpy==1.24.3
opencv-python==4.7.0.72
pyautogui==0.9.54
pillow==9.5.0
matplotlib==3.7.2
python-xlib==0.33
pyscreenshot==3.1 